Refactoring... - Created ROS lib to be used by other packages - Converted executable to lib - Consider adding setters/getters as needed based on BT implementation.
